# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareNear API",
        "description": "CareNear backend API — Plateforme de mise en relation patients / infirmiers à domicile.",
        "contact": {
            "email": "contact@carenear.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Local dev server"
        }
    ],
    "paths": {
...
        "/care-requests/{careRequest}/review": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review for a completed care request",
                "description": "PATIENT only. Only allowed after status is DONE. One review per care request.",
                "operationId": "cfd3b2f294927064c0911de95195ba9f",
                "parameters": [
                    {
                        "name": "careRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Très professionnel et ponctuel.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created"
                    },
                    "403": {
                        "description": "Forbidden — not the patient or not PATIENT role"
                    },
                    "409": {
                        "description": "Review already exists for this request"
                    },
                    "422": {
                        "description": "Request is not DONE or has no nurse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/nurses/{nurseUserId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List reviews for a nurse (public)",
                "description": "No authentication required. Returns paginated reviews.",
                "operationId": "7bb2feaf725949480062dd7a1b224667",
                "parameters": [
                    {
                        "name": "nurseUserId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/nurses/{nurseUserId}/rating": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get aggregated rating for a nurse (public)",
                "description": "No authentication required. Returns average rating and total review count.",
                "operationId": "bec8cdd54a67b8490c4409ca1a0d58fa",
                "parameters": [
                    {
                        "name": "nurseUserId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "nurse_user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "rating_avg": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 4.5
                                                },
                                                "reviews_count": {
                                                    "type": "integer",
                                                    "example": 12
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Sanctum token as Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}